fix: rename homescreen to homepage
