+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css'
-import { HomePage } from './pages/Homescreen/HomePages'
+import { HomePage } from './pages/HomePage/HomePage'
 
 function App() {
   return (
